Address review: extract ProfilePhotoMenu and CameraPicker to separate files
Moves ProfilePhotoMenu and CameraPicker out of ShareProfileView.swift
into their own files for better organization.

Note: These files need to be added to the Xcode project.
