+++ podcasts/Share Profile/ShareProfileView.swift
@@ -706,71 +706,6 @@ struct EditPhotoAndNameView: View {
 
 // MARK: - Shared Components
 
-private struct ProfilePhotoMenu: View {
-    @EnvironmentObject var theme: Theme
-    @ObservedObject var viewModel: ShareProfileViewModel
-
-    var body: some View {
-        VStack(spacing: 12) {
-            Menu {
-                Button {
-                    viewModel.showingPhotoPicker = true
-                } label: {
-                    Label(L10n.shareProfileChoosePhoto, systemImage: "photo.on.rectangle")
-                }
-
-                Button {
-                    viewModel.showingCamera = true
-                } label: {
-                    Label(L10n.shareProfileTakePhoto, systemImage: "camera")
-                }
-
-                if viewModel.profilePhoto != nil {
-                    Divider()
-
-                    Button(role: .destructive) {
-                        viewModel.removePhoto()
-                    } label: {
-                        Label(L10n.shareProfileRemovePhoto, systemImage: "trash")
-                    }
-                }
-            } label: {
-                ZStack(alignment: .bottomTrailing) {
-                    if let photo = viewModel.profilePhoto {
-                        Image(uiImage: photo)
-                            .resizable()
-                            .aspectRatio(contentMode: .fill)
-                            .frame(width: 134, height: 134)
-                            .clipShape(Circle())
-                    } else {
-                        ProfileImage(email: viewModel.email)
-                            .frame(width: 134, height: 134)
-                            .clipShape(Circle())
-                    }
-
-                    Image("folder-edit")
-                        .renderingMode(.template)
-                        .resizable()
-                        .aspectRatio(contentMode: .fit)
-                        .foregroundColor(theme.primaryInteractive02)
-                        .frame(width: 24, height: 24)
-                        .frame(width: 36, height: 36)
-                        .background(Circle().fill(theme.primaryInteractive01))
-                        .clipShape(Circle())
-                        .overlay(Circle().stroke(theme.primaryUi01, lineWidth: 3))
-                }
-            }
-        }
-        .photosPicker(isPresented: $viewModel.showingPhotoPicker, selection: $viewModel.selectedPhotoItem, matching: .images)
-        .fullScreenCover(isPresented: $viewModel.showingCamera) {
-            CameraPicker { image in
-                viewModel.profilePhoto = image
-            }
-            .ignoresSafeArea()
-        }
-    }
-}
-
 private struct DisplayNameField: View {
     @EnvironmentObject var theme: Theme
     @Binding var displayName: String
@@ -822,44 +757,6 @@ private struct BottomGradientContainer<Content: View>: View {
 
             content
                 .background(theme.primaryUi01)
-        }
-    }
-}
-
-// MARK: - Camera Picker
-
-struct CameraPicker: UIViewControllerRepresentable {
-    var onImagePicked: (UIImage) -> Void
-
-    func makeUIViewController(context: Context) -> UIImagePickerController {
-        let picker = UIImagePickerController()
-        picker.sourceType = .camera
-        picker.delegate = context.coordinator
-        return picker
-    }
-
-    func updateUIViewController(_ uiViewController: UIImagePickerController, context: Context) {}
-
-    func makeCoordinator() -> Coordinator {
-        Coordinator(onImagePicked: onImagePicked)
-    }
-
-    class Coordinator: NSObject, UIImagePickerControllerDelegate, UINavigationControllerDelegate {
-        let onImagePicked: (UIImage) -> Void
-
-        init(onImagePicked: @escaping (UIImage) -> Void) {
-            self.onImagePicked = onImagePicked
-        }
-
-        func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey: Any]) {
-            if let image = info[.originalImage] as? UIImage {
-                onImagePicked(image)
-            }
-            picker.dismiss(animated: true)
-        }
-
-        func imagePickerControllerDidCancel(_ picker: UIImagePickerController) {
-            picker.dismiss(animated: true)
         }
     }
 }
